pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monthly-limit-order-review-system"
version = "0.1.0"
description = "CLI for monthly limit-order review prompt generation and review ingestion"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2",
  "PyYAML>=6.0",
  "yfinance>=0.2.54",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

